requirements.txt
asgiref==3.4.1
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-cors-headers==3.10.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.3.2
PyJWT==2.2.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.2
twilio==7.2.0
uritemplate==4.1.1
urllib3==1.26.7